airflow/requirements.txt
pandas==2.0.3
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3